% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rd_num.R
\name{get_rd_num}
\alias{get_rd_num}
\title{Get the Round Number}
\usage{
get_rd_num(maxRounds = NULL)
}
\arguments{
\item{maxRounds:}{The maximum number of rounds in the tournament (numeric)}
}
\value{
The user inputed round number
}
\description{
This function asks a user to input the round number for a completed
debate round. The function checks to see if the input is less than
the maximum number of rounds in the tournament, greater than 0, and
that all debates in the previous round has been completed. If any of
these conditions are not satisfied, it will ask the user to input a
different round number. To exit the loop, the user can type
\kbd{quit} and the function will return NULL.
}
\examples{
get_rd_num()
get_rd_num(maxRounds=5)
}
